added customer in program.cs

+++ SD_Assign_Mediator_Pattern_Design/AbstractAirline.cs
@@ -8,17 +8,13 @@ namespace SD_Assign_Mediator_Pattern_Design
     public abstract class AbstractAirLine
     {
        //public bool FullBooked { get; set; }
-
-        
-
-
        
         public int Duration = 0;
         public int stops = 0;
         public static double Discount { get; set; }
         public string NameOfAirline = "";
         public double Rate = 4.4;
-        public static double Price = 0.0;
+        public static int Price = 0;
         public bool Available = true;
         
         public abstract void CancelFlight();
